Refactor: Organize code and simplify structure
Moved the `restore_blob_with_sha` function from `tree.rs` to `packset.rs` and made it a method of a new `PackSet` struct.

Created a new `commit.rs` module and moved the `Commit` struct and its implementation from `tree.rs` to the new module.

Refactored `lib.rs` to use the new `commit.rs` module.

Refactored `tree.rs` to remove the `Commit` struct and its implementation.

Fixed all warnings and failing tests.
