+++ arq/src/arq7.rs
@@ -757,105 +757,6 @@ impl BlobLoc {
         })
     }
 
-    /// Fallback parsing method for compatibility
-    fn from_binary_reader_fallback<R: binary::ArqBinaryReader>(reader: &mut R) -> Result<Self> {
-        let blob_identifier = reader.read_arq_string_required()?;
-        let is_packed = reader.read_arq_bool()?;
-        let is_large_pack_binary = reader.read_arq_bool()?;
-
-        // Simplified path recovery
-        let relative_path = match reader.read_arq_string() {
-            Ok(Some(path)) if Self::is_valid_path(&path) => path,
-            Ok(Some(_)) | Ok(None) => {
-                // Try simple recovery
-                Self::try_simple_path_recovery(reader)
-                    .unwrap_or(None)
-                    .unwrap_or_default()
-            }
-            Err(_) => String::new(),
-        };
-
-        let offset = reader.read_arq_u64().unwrap_or(0);
-        let length = reader.read_arq_u64().unwrap_or(0);
-
-        // Validate offset and length to prevent abnormal values
-        let safe_offset = if offset > 1_000_000_000_000 {
-            0
-        } else {
-            offset
-        };
-        let safe_length = if length > 1_000_000_000_000 {
-            0
-        } else {
-            length
-        };
-
-        let stretch_encryption_key = reader.read_arq_bool().unwrap_or(true);
-        let compression_type = reader.read_arq_u32().unwrap_or(2); // Default to LZ4
-
-        Ok(BlobLoc {
-            blob_identifier,
-            is_packed,
-            is_large_pack: Some(is_large_pack_binary),
-            relative_path,
-            offset: safe_offset,
-            length: safe_length,
-            stretch_encryption_key,
-            compression_type,
-        })
-    }
-
-    /// Simple path recovery without consuming too much data
-    fn try_simple_path_recovery<R: binary::ArqBinaryReader>(
-        reader: &mut R,
-    ) -> Result<Option<String>> {
-        // Try one recovery attempt - read a string and validate it
-        match reader.read_arq_string() {
-            Ok(Some(potential_path)) if Self::is_valid_path(&potential_path) => {
-                Ok(Some(potential_path))
-            }
-            _ => Ok(None),
-        }
-    }
-
-    /// Validate if a string looks like a reasonable file path
-    fn is_valid_path(path: &str) -> bool {
-        // Basic validation for path-like strings
-        if path.is_empty() || path.len() > 4096 {
-            return false;
-        }
-
-        // Check for typical backup path patterns
-        let has_backup_patterns = path.contains("treepacks")
-            || path.contains("blobpacks")
-            || path.contains("largeblobpacks")
-            || path.contains(".pack")
-            || path.contains("standardobjects")
-            || path.contains("blob");
-
-        // Should start with / for absolute paths or contain backup patterns
-        if !path.starts_with('/') && !has_backup_patterns {
-            return false;
-        }
-
-        // Should contain mostly printable ASCII characters
-        // Allow common path characters and avoid control characters
-        let has_valid_chars = path.chars().all(|c| {
-            !c.is_control()
-                && (c.is_ascii_alphanumeric()
-                    || c == '/'
-                    || c == '-'
-                    || c == '_'
-                    || c == '.'
-                    || c == ' '
-                    || c == '('
-                    || c == ')'
-                    || c == ':'
-                    || c == '\\')
-        });
-
-        has_valid_chars
-    }
 }
 
 /// Conversion from arq7::BlobLoc to blob_location::BlobLoc
