+++ arq/src/packset.rs
@@ -29,11 +29,84 @@ use byteorder::{NetworkEndian, ReadBytesExt};
 use std;
 use std::io::{BufRead, Cursor, Seek, SeekFrom};
 
+use std::fs::{self, File};
+use std::io::BufReader;
+use std::path::{Path, PathBuf};
+
 use crate::compression::CompressionType;
 use crate::error::Result;
 use crate::object_encryption::{calculate_sha1sum, EncryptedObject};
 use crate::type_utils::ArqRead;
 use crate::utils::convert_to_hex_string;
+
+pub struct PackSet {
+    path: PathBuf,
+}
+
+impl PackSet {
+    pub fn new(path: &Path) -> Self {
+        Self {
+            path: path.to_path_buf(),
+        }
+    }
+
+    // TODO: Do this better - don't read all files multiple times
+    pub fn restore_blob_with_sha(
+        &self,
+        sha: &str,
+        keyset: &crate::arq7::EncryptedKeySet,
+    ) -> Result<Option<Vec<u8>>> {
+        for entry_result in fs::read_dir(&self.path)? {
+            let entry = entry_result?;
+
+            let fname = entry.file_name();
+            let fname_str = match fname.to_str() {
+                Some(s) => s,
+                None => {
+                    continue;
+                }
+            };
+
+            if fname_str.ends_with(".index") {
+                let index_path = entry.path();
+                let mut reader = get_file_reader_for_restore(&index_path)
+                    .map_err(|e| crate::error::Error::IoError(e))?;
+
+                let index = PackIndex::new(&mut reader)?;
+
+                for obj in index.objects {
+                    if obj.sha1 == sha {
+                        let pack_path = index_path.with_extension("pack");
+                        if !pack_path.exists() {
+                            continue;
+                        }
+                        let mut pack_reader = get_file_reader_for_restore(&pack_path)
+                            .map_err(|e| crate::error::Error::IoError(e))?;
+
+                        pack_reader
+                            .seek(SeekFrom::Start(obj.offset as u64))
+                            .map_err(|e| crate::error::Error::IoError(e))?;
+
+                        let pack = PackObject::new(&mut pack_reader)?;
+
+                        match pack.data.decrypt(&keyset.encryption_key) {
+                            Ok(data) => return Ok(Some(data)),
+                            Err(e) => return Err(e),
+                        }
+                    }
+                }
+            }
+        }
+        Ok(None)
+    }
+}
+
+// Helper function to get a BufReader for a file, returning std::io::Error on failure.
+// Uses std::io::BufReader which is already imported at the top of the file.
+fn get_file_reader_for_restore(path: &Path) -> std::io::Result<BufReader<File>> {
+    let file = File::open(path)?;
+    Ok(BufReader::new(file))
+}
 
 ///Pack File Format
 ///----------------
