+++ arq/src/tree.rs
@@ -19,78 +19,15 @@
 //! All commits, trees and blobs are typically stored as EncryptedObjects in Arq backups.
 use std;
 use std::collections::HashMap;
-use std::io::{BufReader, Seek, SeekFrom};
+use std::io::BufReader;
 
 use byteorder::ReadBytesExt;
-use chrono::{DateTime, Utc};
 
 use crate::arq7::binary::ArqBinaryReader;
 use crate::blob;
 use crate::compression::CompressionType;
 use crate::error::Result;
 use crate::type_utils::ArqRead;
-
-use crate::packset::{PackIndex, PackObject};
-use std::fs::{self, File};
-
-use std::path::Path;
-
-// Helper function to get a BufReader for a file, returning std::io::Error on failure.
-// Uses std::io::BufReader which is already imported at the top of the file.
-fn get_file_reader_for_restore(path: &Path) -> std::io::Result<BufReader<File>> {
-    let file = File::open(path)?;
-    Ok(BufReader::new(file))
-}
-
-// TODO: Do this better - don't read all files multiple times
-pub fn restore_blob_with_sha(
-    path: &Path,
-    sha: &str,
-    keyset: &crate::arq7::EncryptedKeySet,
-) -> Result<Option<Vec<u8>>> {
-    for entry_result in fs::read_dir(path)? {
-        let entry = entry_result?;
-
-        let fname = entry.file_name();
-        let fname_str = match fname.to_str() {
-            Some(s) => s,
-            None => {
-                continue;
-            }
-        };
-
-        if fname_str.ends_with(".index") {
-            let index_path = entry.path();
-            let mut reader = get_file_reader_for_restore(&index_path)
-                .map_err(|e| crate::error::Error::IoError(e))?;
-
-            let index = PackIndex::new(&mut reader)?;
-
-            for obj in index.objects {
-                if obj.sha1 == sha {
-                    let pack_path = index_path.with_extension("pack");
-                    if !pack_path.exists() {
-                        continue;
-                    }
-                    let mut pack_reader = get_file_reader_for_restore(&pack_path)
-                        .map_err(|e| crate::error::Error::IoError(e))?;
-
-                    pack_reader
-                        .seek(SeekFrom::Start(obj.offset as u64))
-                        .map_err(|e| crate::error::Error::IoError(e))?;
-
-                    let pack = PackObject::new(&mut pack_reader)?;
-
-                    match pack.data.decrypt(&keyset.encryption_key) {
-                        Ok(data) => return Ok(Some(data)),
-                        Err(e) => return Err(e),
-                    }
-                }
-            }
-        }
-    }
-    Ok(None)
-}
 
 /// Tree
 ///
@@ -500,153 +437,6 @@ impl Tree {
             st_blksize: 0,                                  // Default
             missing_nodes: Vec::new(), // Default (specific to Arq5 tree parsing)
             nodes: child_nodes_map,
-        })
-    }
-}
-
-pub type ParentCommits = HashMap<String, bool>;
-pub type FailedFile = (String, String);
-
-/// Commit
-///
-/// A "commit" contains the following bytes:
-///
-/// ```ascii
-///     43 6f 6d 6d 69 74 56 30 31 32      "CommitV012"
-///     [String:"<author>"]
-///     [String:"<comment>"]
-///     [UInt64:num_parent_commits]        (this is always 0 or 1)
-///     (
-///         [String:parent_commit_sha1] /* can't be null */
-///         [Bool:parent_commit_encryption_key_stretched]] /* present for Commit version >= 4 */
-///     )   /* repeat num_parent_commits times */
-///     [String:tree_sha1]] /* can't be null */
-///     [Bool:tree_encryption_key_stretched]] /* present for Commit version >= 4 */
-///     [Bool:tree_is_compressed] /* present for Commit version 8 and 9 only; indicates Gzip compression or none */
-///     [CompressionType:tree_compression_type] /* present for Commit version >= 10 */
-///
-///     [String:"file://<hostname><path_to_folder>"]
-///     [String:"<merge_common_ancestor_sha1>"] /* only present for Commit version 7 or *older* (was never used) */
-///     [Bool:is_merge_common_ancestor_encryption_key_stretched] /* only present for Commit version 4 to 7 */
-///     [Date:creation_date]
-///     [UInt64:num_failed_files] /* only present for Commit version 3 or later */
-///     (
-///         [String:"<relative_path>"] /* only present for Commit version 3 or later */
-///         [String:"<error_message>"] /* only present for Commit version 3 or later */
-///     )   /* repeat num_failed_files times */
-///     [Bool:has_missing_nodes] /* only present for Commit version 8 or later */
-///     [Bool:is_complete] /* only present for Commit version 9 or later */
-///     [Data:config_plist_xml] /* a copy of the XML file as described above */
-///     [String:arq_version] /* the version of the Arq app that created this Commit */
-/// ```
-///
-/// The SHA1 of the most recent Commit is stored in
-/// `/<computer_uuid>/bucketdata/<folder_uuid>/refs/heads/master` appended with a "Y" for
-/// historical reasons.
-///
-/// In addition, Arq writes a file in
-/// `/<computer_uuid>/bucketdata/<folder_uuid>/refs/logs/master` each time a new Commit is
-/// created (the filename is a timestamp). It's a plist containing the previous and current
-/// Commit SHA1s, the SHA1 of the pack file containing the new Commit, and whether the new
-/// Commit is a "rewrite" (because the user deleted a backup record for instance).
-pub struct Commit {
-    pub version: u32,
-    pub author: String,
-    pub comment: String,
-    pub parent_commits: HashMap<String, bool>,
-    pub tree_sha1: String,
-    pub tree_encryption_key_stretched: bool,
-    pub tree_compression_type: CompressionType,
-    pub folder_path: String,
-    pub creation_date: Option<DateTime<Utc>>,
-    pub failed_files: Vec<(String, String)>,
-    pub has_missing_nodes: bool,
-    pub is_complete: bool,
-    pub config_plist_xml: Vec<u8>,
-    pub arq_version: String,
-}
-
-impl Commit {
-    pub fn is_commit(content: &[u8]) -> bool {
-        content[..10] == [67, 111, 109, 109, 105, 116, 86, 48, 49, 50] // CommitV012
-    }
-
-    pub fn new<R: ArqRead>(mut reader: R) -> Result<Commit> {
-        let header = reader.read_bytes(10)?;
-        assert_eq!(header[..7], [67, 111, 109, 109, 105, 116, 86]); // CommitV
-        let version = std::str::from_utf8(&header[7..])?.parse::<u32>()?;
-
-        let author = reader.read_arq_string()?;
-        let comment = reader.read_arq_string()?;
-
-        let mut num_parent_commits = reader.read_arq_u64()?;
-        assert!(num_parent_commits == 0 || num_parent_commits == 1);
-
-        let mut parent_commits: HashMap<String, bool> = HashMap::new();
-        while num_parent_commits > 0 {
-            let sha1 = reader.read_arq_string()?;
-            let encryption_key_stretched = reader.read_arq_bool()?;
-
-            parent_commits.insert(sha1, encryption_key_stretched);
-            num_parent_commits -= 1;
-        }
-
-        let tree_sha1 = reader.read_arq_string()?;
-        let tree_encryption_key_stretched = reader.read_arq_bool()?;
-        let tree_compression_type = reader.read_arq_compression_type()?;
-        let folder_path = reader.read_arq_string()?;
-
-        // Read and convert creation_date
-        let parsed_creation_date: Option<DateTime<Utc>>;
-        let present_byte = reader.read_bytes(1)?;
-        if present_byte[0] == 0x01 {
-            let milliseconds_since_epoch = reader.read_arq_u64()?;
-            if milliseconds_since_epoch == 0 {
-                parsed_creation_date = None;
-            } else {
-                parsed_creation_date =
-                    DateTime::from_timestamp_millis(milliseconds_since_epoch as i64);
-                if parsed_creation_date.is_none() {
-                    return Err(crate::error::Error::InvalidFormat(format!(
-                        "Invalid timestamp for commit creation_date: {}ms",
-                        milliseconds_since_epoch
-                    )));
-                }
-            }
-        } else {
-            parsed_creation_date = None;
-        }
-
-        let mut num_failed_files = reader.read_arq_u64()?;
-        let mut failed_files: Vec<(String, String)> = Vec::new();
-        while num_failed_files > 0 {
-            let relative_path = reader.read_arq_string()?;
-            let error_message = reader.read_arq_string()?;
-
-            failed_files.push((relative_path, error_message));
-            num_failed_files -= 1;
-        }
-
-        let has_missing_nodes = reader.read_arq_bool()?;
-        let is_complete = reader.read_arq_bool()?;
-        let config_plist_xml = reader.read_arq_data()?;
-        let arq_version = reader.read_arq_string()?;
-
-        Ok(Commit {
-            version,
-            author,
-            comment,
-            parent_commits,
-            tree_sha1,
-            tree_encryption_key_stretched,
-            tree_compression_type,
-            folder_path,
-            creation_date: parsed_creation_date,
-            failed_files,
-            has_missing_nodes,
-            is_complete,
-            config_plist_xml,
-            arq_version,
         })
     }
 }
