+++ arq/tests/arq7_test.rs
@@ -5,10 +5,10 @@ use std::io::Cursor;
 use std::path::Path;
 
 const ARQ7_TEST_DATA_DIR_NOT_ENCRYPTED: &str =
-    "/Users/developer/Projects/2025-06-arq/arq/tests/arq_storage_location/2E7BB0B6-BE5B-4A86-9E51-10FE730E1104";
+    "./tests/arq_storage_location/2E7BB0B6-BE5B-4A86-9E51-10FE730E1104";
 
 const ARQ7_TEST_DATA_DIR_ENCRYPTED: &str =
-    "/Users/developer/Projects/2025-06-arq/arq/tests/arq_storage_location/D1154AC6-01EB-41FE-B115-114464350B92";
+    "./tests/arq_storage_location/D1154AC6-01EB-41FE-B115-114464350B92";
 // const ARQ7_TEST_DATA_DIR_NOT_ENCRYPTED: &str =
 //     "/Users/developer/Projects/2025-06-arq/arq/tests/arq_storage_location/FD5575D9-B7E1-43D9-B29C-B54ACC9BC2A9";
